Started work on the actual server.
