+++ server/ropsite/scheduler/templates/scheduler/index.html
@@ -1,0 +1,3 @@
+{% load staticfiles %}
+<h1>Velkommen til RevyØveplanlægningsProgrammelet(ROP)!!</h1>
+<link rel="stylesheet" type="text/css" href="{% static 'polls/style.css' %}" />
