+++ server/ropsite/scheduler/views.py
@@ -1,0 +1,8 @@
+from django.shortcuts import render
+from django.http import HttpResponse
+from django.views import generic
+
+# The index page
+
+def index(request):
+    return HttpResponse("Hello, world. You're at the polls index.")
